Update QR scanner and improve tranform algorithm
The QR scanner is now much more accurate. It's using a pre-compiled
scanner from Daniel Beer (https://github.com/dlbeer/quirc.git). There
are some issues with getting this to work with JPEG images however, so
at the moment all input images are being converted to PNG (this is quite
a slow process and a takes a majority of the processing time).